Coupon permissions

+++ src/Http/Controllers/Cp/CouponController.php
@@ -3,6 +3,7 @@
 namespace Damcclean\Commerce\Http\Controllers\Cp;
 
 use Damcclean\Commerce\Facades\Coupon;
+use Damcclean\Commerce\Policies\CouponPolicy;
 use Illuminate\Http\Request;
 use Statamic\Facades\Blueprint;
 use Statamic\Http\Controllers\CP\CpController;
@@ -11,6 +12,8 @@ class CouponController extends CpController
 {
     public function index()
     {
+        $this->authorize('view', CouponPolicy::class);
+
         return view('commerce::cp.coupons.index', [
             'coupons' => Coupon::all()
         ]);
@@ -18,6 +21,8 @@ class CouponController extends CpController
 
     public function create()
     {
+        $this->authorize('create', CouponPolicy::class);
+
         $blueprint = Blueprint::find('coupon');
 
         $fields = $blueprint->fields();
@@ -33,6 +38,8 @@ class CouponController extends CpController
 
     public function store(Request $request)
     {
+        $this->authorize('create', CouponPolicy::class);
+
         $validated = []; // WIP
 
         $coupon = Coupon::save($request->all());
@@ -44,6 +51,8 @@ class CouponController extends CpController
 
     public function edit($product)
     {
+        $this->authorize('edit', CouponPolicy::class);
+
         $coupon = Coupon::find($product);
 
         $blueprint = Blueprint::find('coupon');
@@ -61,6 +70,8 @@ class CouponController extends CpController
 
     public function update(Request $request, $coupon)
     {
+        $this->authorize('edit', CouponPolicy::class);
+
         $validated = []; // wip
 
         $coupon = Coupon::update(Coupon::find($coupon)['slug'], $request->all());
@@ -74,6 +85,8 @@ class CouponController extends CpController
 
     public function destroy($coupon)
     {
+        $this->authorize('delete', CouponPolicy::class);
+
         $coupon = Coupon::delete(Coupon::find($coupon)['slug']);
 
         return redirect(cp_route('coupons.index'));
